move timespans to constants
