+++ src/shared/constants.ts
@@ -22,3 +22,21 @@ export const enum Timespans {
   FOUR_WEEKS_IN_SECONDS = 60 * 60 * 24 * 7 * 4,
   ONE_YEAR_IN_SECONDS = 60 * 60 * 24 * 365,
 }
+
+export const enum MuteDuration {
+  OFF = Timespans.ZERO_SECONDS,
+  ONE_HOUR = Timespans.ONE_HOUR_IN_SECONDS,
+  TWO_HOURS = Timespans.TWO_HOURS_IN_SECONDS,
+  ONE_DAY = Timespans.ONE_DAY_IN_SECONDS,
+  SEVEN_DAYS = Timespans.ONE_WEEK_IN_SECONDS,
+  FOREVER = -1,
+}
+
+export const enum AutodeleteDuration {
+  OFF = Timespans.ZERO_SECONDS,
+  ONE_HOUR = Timespans.ONE_HOUR_IN_SECONDS,
+  ONE_DAY = Timespans.ONE_DAY_IN_SECONDS,
+  ONE_WEEK = Timespans.ONE_WEEK_IN_SECONDS,
+  FOUR_WEEKS = Timespans.FOUR_WEEKS_IN_SECONDS,
+  ONE_YEAR = Timespans.ONE_YEAR_IN_SECONDS,
+}
