+++ src/shared/shared-types.d.ts
@@ -167,23 +167,3 @@ export type Theme = {
   description: string
   address: string
 }
-
-import { Timespans } from './constants'
-
-export const enum MuteDuration {
-  OFF = Timespans.ZERO_SECONDS,
-  ONE_HOUR = Timespans.ONE_HOUR_IN_SECONDS,
-  TWO_HOURS = Timespans.TWO_HOURS_IN_SECONDS,
-  ONE_DAY = Timespans.ONE_DAY_IN_SECONDS,
-  SEVEN_DAYS = Timespans.ONE_WEEK_IN_SECONDS,
-  FOREVER = -1,
-}
-
-export enum AutodeleteDuration {
-  OFF = Timespans.ZERO_SECONDS,
-  ONE_HOUR = Timespans.ONE_HOUR_IN_SECONDS,
-  ONE_DAY = Timespans.ONE_DAY_IN_SECONDS,
-  ONE_WEEK = Timespans.ONE_WEEK_IN_SECONDS,
-  FOUR_WEEKS = Timespans.FOUR_WEEKS_IN_SECONDS,
-  ONE_YEAR = Timespans.ONE_YEAR_IN_SECONDS,
-}
